+++ utils/gnuplot_i.h
@@ -257,7 +257,7 @@ void gnuplot_resetplot(gnuplot_ctrl * h);
   @endcode
  */
 /*--------------------------------------------------------------------------*/
-void gnuplot_plot_x(gnuplot_ctrl * handle, double * d, int n, char * title);
+void gnuplot_plot_x(gnuplot_ctrl * handle, float * d, int n, char * title);
 
 /*-------------------------------------------------------------------------*/
 /**
@@ -292,8 +292,8 @@ void gnuplot_plot_x(gnuplot_ctrl * handle, double * d, int n, char * title);
 /*--------------------------------------------------------------------------*/
 void gnuplot_plot_xy(
     gnuplot_ctrl    *   handle,
-    double          *   x,
-    double          *   y,
+    float          *   x,
+    float          *   y,
     int                 n,
     char            *   title
 ) ;
@@ -324,8 +324,8 @@ void gnuplot_plot_once(
     char    *   style,
     char    *   label_x,
     char    *   label_y,
-    double  *   x,
-    double  *   y,
+    float  *   x,
+    float  *   y,
     int         n
 );
 
@@ -356,8 +356,8 @@ void gnuplot_plot_once(
 /*--------------------------------------------------------------------------*/
 void gnuplot_plot_slope(
     gnuplot_ctrl    *   handle,
-    double              a,
-    double              b,
+    float              a,
+    float              b,
     char            *   title
 ) ;
